backend/requirements.txt
Django==4.2.23
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djoser==2.3.1
django-import-export
gunicorn==20.1.0 
Pillow==11.3.0
psycopg2-binary==2.9.6